requirements.txt
python-dotenv
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
certifi==2025.6.15
click==8.2.1
fastapi==0.115.14
granian==2.4.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
platformdirs==4.3.8
psycopg[binary]
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
redis==6.2.0
reflex==0.8.0
reflex-hosting-cli==0.1.51
rich==14.0.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
watchfiles==1.1.0
wrapt==1.17.2
wsproto==1.2.0
